Update Form1.cs
2020-03-29
Added text box for UUID, so it can be entered manually. If this is left empty, a new UUID will be generated.
(As before, if the envelope details are copied from a previous sequence, using the "copy old envelope" button, the UUID of that sequence will be autmatically filled in.)

The attributes for the following Modules are no longer copied from the file path, but can now be read from a text file:
2.3.S substance and manufacturer
2.3.P product-name, dosageform and manufacturer
3.2.S substance and manufacturer
3.2.P product-name, dosageform and manufacturer
3.2.P.4 excipient
3.2.A.1 product-name, dosageform, manufacturer and substance
3.2.A.2 product-name, dosageform, manufacturer and substance

The text file should be named "attributes.txt" and be saved under the top level folder (i.e. on the same level as the sequence folders, not inside a sequence folder). A template text file is included. For further information, see the instructions.